feat 금액 입력창 컴포넌트화
